refactor(plugins): name the file after the standard it scores
cisa.py sat next to cisa_2025.py, so the pair read as "the CISA plugin"
and "an older one" rather than as two editions of the same standard. The
2025 draft and the 2026 final are both current, for different reasons, so
neither gets to be the unqualified name.

Nothing about the plugin's identity moves. PLUGIN_NAME is an explicit
constant, so stored AssessmentRun.plugin_name rows are untouched, and
_register passes plugin_class_path through update_or_create keyed on the
name, so the RegisteredPlugin row rewrites itself on the next boot. No
migration.

test_cisa_plugin.py becomes test_cisa_2026_plugin.py for the same reason
it already has a test_cisa_2025_plugin.py sibling.
